refactor(app): decouple Function and DisplayName
